Fixing issue on Transaction and Payment

+++ Rekord/Core/Entities/PaymentModel.swift
@@ -10,11 +10,11 @@ class PaymentModel {
     var idTransaction: String?
     var idUser: String?
     var createdDate: String?
-    var amount: Float?
+    var amount: Double?
     var document: String?
     var airtableId: String?
     
-    init(idPayment: String, idTransaction: String, idUser: String, createdDate: String, amount: Float, document: String, airtableId: String) {
+    init(idPayment: String, idTransaction: String, idUser: String, createdDate: String, amount: Double, document: String, airtableId: String) {
         self.idPayment = idPayment
         self.idTransaction = idTransaction
         self.idUser = idUser
